1. [board repo/serivce/controller.ts] board remove 로직 유저 권환 체크 및 id 조회 2. [update-board.dto.ts, create-board.dto.ts] : property 추가 및 조건 수정

+++ src/study/board/board.service.ts
@@ -1,6 +1,7 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { HttpException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CommentRepository } from 'src/comment/comment.repository';
+import { UserEntity } from 'src/user/entities/user.entity';
 import { UserService } from 'src/user/user.service';
 // import { StudyBoardEntity } from 'src/study/board/entity/board.entity';
 import { Repository } from 'typeorm';
@@ -56,7 +57,20 @@ export class BoardService {
     return this.studyBoardRepo.updateBoard(id, body);
   }
 
-  remove(id: number) {
+  async remove(id: number, userId?: number) {
+    const boardData = await this.studyBoardRepo.findBoard(id);
+
+    if (!userId) throw new HttpException('Not Access Authorization ', 403);
+
+    if (boardData) {
+      //@ts-ignore
+      if (boardData.writer.id !== userId) {
+        // # 해당 게시판에 대해 삭제 권한이 없을 경우
+        throw new HttpException('Not Access Authorization ', 403);
+      }
+    } else {
+      throw new HttpException('No Board', 403);
+    }
     return this.studyBoardRepo.removeBoard(id);
   }
 }
